auth and syntax fixes
