+++ middleware.ts
@@ -1,8 +1,50 @@
-import { type NextRequest } from "next/server"
-import { updateSession } from "@/lib/supabase/middleware"
+import { createServerClient, type CookieOptions } from "@supabase/ssr"
+import { type NextRequest, NextResponse } from "next/server"
 
 export async function middleware(request: NextRequest) {
-  return await updateSession(request)
+  // Create an unmodified response
+  let response = NextResponse.next({
+    request: {
+      headers: request.headers
+    }
+  })
+
+  const supabase = createServerClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      cookies: {
+        get(name: string) {
+          return request.cookies.get(name)?.value
+        },
+        set(name: string, value: string, options: CookieOptions) {
+          // If the cookie is updated, update the cookies for the request and response
+          request.cookies.set({ name, value, ...options })
+          response = NextResponse.next({
+            request: {
+              headers: request.headers
+            }
+          })
+          response.cookies.set({ name, value, ...options })
+        },
+        remove(name: string, options: CookieOptions) {
+          // If the cookie is removed, update the cookies for the request and response
+          request.cookies.set({ name, value: "", ...options })
+          response = NextResponse.next({
+            request: {
+              headers: request.headers
+            }
+          })
+          response.cookies.set({ name, value: "", ...options })
+        }
+      }
+    }
+  )
+
+  // This will refresh the session if expired - required for Server Components
+  await supabase.auth.getUser()
+
+  return response
 }
 
 export const config = {
